backend/requirements.txt
fastapi[standard]
uvicorn
boto3
azure-identity
azure-mgmt-costmanagement
psycopg2-binary
sqlalchemy
python-jose
passlib
cryptography
python-dotenv